Create README.md
# CaptureAndSaveImage
With the Help of ImagePicker Library got by searching in Google. The Current Project is Developed.
This project deal with the Capturing/Browesing an Image and Set It to a ImageView.
And in this project you can find the Code to Store the Image File in an Specific Directory for Below and Above Android Q Versions.
Image will be Stored in the "/Pictures/SaveImages" Folder. 


............Happy Coding!!!!